Suggest usage of sudo when root access is needed.
Also be explicit about requirement for the NFS/iscsi server.

A user has to configured both the NFS/iSCSI server and the kubernetes
nodes it is clearer if the configuration clarifies which "set of
servers" required each type of configuration.  This tiny change
makes that more explicit and the use of sudo is because a lot of
kubernetes commands do not require root, so if root is required it is
good to be explicit.